+++ config/ai.go
@@ -3,6 +3,7 @@ package config
 const (
 	openai_key         = "OPENAI_KEY"
 	openai_model       = "OPENAI_MODEL"
+	openai_base_url    = "OPENAI_BASE_URL"
 	openai_proxy       = "OPENAI_PROXY"
 	openai_temperature = "OPENAI_TEMPERATURE"
 	openai_max_tokens  = "OPENAI_MAX_TOKENS"
@@ -11,6 +12,7 @@ const (
 type AiConfig struct {
 	key         string
 	model       string
+	baseURL     string
 	proxy       string
 	temperature float64
 	maxTokens   int
@@ -22,6 +24,10 @@ func (c AiConfig) GetKey() string {
 
 func (c AiConfig) GetModel() string {
 	return c.model
+}
+
+func (c AiConfig) GetBaseURL() string {
+	return c.baseURL
 }
 
 func (c AiConfig) GetProxy() string {